🔨 refactor: fsd public api 사용하도록 변경
