+++ src/features/history/index.ts
@@ -1,3 +1,9 @@
 export { useBookNavigation } from './model/useBookNavigation';
-export { INDEX_LIST, PAGE_SIDE } from './model/constants';
-export type { IndexItem, PageSide, FlipDirection } from './model/types';
+export { useBookCoverState } from './model/useBookCoverState';
+export { INDEX_LIST, PAGE_SIDE, RAPID_FLIP_DURATION } from './model/constants';
+export type {
+  IndexItem,
+  PageSide,
+  FlipDirection,
+  BookState,
+} from './model/types';
