+++ src/shared/lib/analytics/index.ts
@@ -1,0 +1,1 @@
+export { useGoogleAnalytics } from './useGoogleAnalytics';
