+++ src/shared/lib/breakpoint/index.ts
@@ -1,0 +1,6 @@
+export {
+  useBreakpoint,
+  BREAKPOINT_MOBILE_MAX,
+  BREAKPOINT_TABLET_MAX,
+} from './useBreakpoint';
+export type { Breakpoint } from './useBreakpoint';
