+++ src/shared/lib/scroll/index.ts
@@ -1,0 +1,1 @@
+export { smoothScrollTo } from './smoothScrollTo';
