+++ src/shared/lib/useIntersectionAnimation/index.ts
@@ -1,0 +1,1 @@
+export { useIntersectionAnimation } from './useIntersectionAnimation';
