+++ src/shared/lib/useSlideGesture/index.ts
@@ -1,0 +1,1 @@
+export { useSlideGesture } from './useSlideGesture';
